Extend CLI to allow consuming prebuilt app
I will use both these flags together with our Nix build pipeline.

We will still read the config from the Cargo.toml, because that is the
metadata source of truth, but the ELF exe will already be prebuilt and
in a read-only location. We'll therefore have the json and hex
side-by-side, and then move then together to the output location.

Now that the CLI is more complex, I took the opportunity to convert it
to the declarative style of Clap --- it's really nice!